refactored DatabaseSearchPage
