+++ client/frontend/src/pages/DatabaseSearchPage.vue
@@ -242,8 +242,8 @@
 
 <script>
 import { useTheme } from 'vuetify'
-import gql from 'graphql-tag'
 import DatabaseSearchFilters from '@/components/DatabaseSearchFilters.vue'
+import { buildQraphQLQuery } from '@/libs/graphql-query.js'
 // import { buildQraphQLQuery } from '@/libs/graphql-query'
 // for Composition API apollo provider:
 // import { useApolloClient } from '@vue/apollo-composable'
@@ -256,9 +256,6 @@ export default {
   components: { DatabaseSearchFilters },
   data: () => ({
     tableIsLoading: false,
-    grouped: 'grouped',
-    repeated: 'repeated',
-    tableStyle: 'repeated',
     exportOptions: [
       { title: 'the search result', icon: 'mdi-table' },
       { title: 'the whole database', icon: 'mdi-database' }
@@ -270,172 +267,13 @@ export default {
     selectedAttrs: []
   }),
   methods: {
-    getFields(model) {
-      return this.selectedAttrs.filter((attr) => attr.model === model).map((attr) => attr.field)
-    },
-    buildQraphQLQuery() {
-      let novelFoodQueryPart = ''
-      let questioinsQueryPart = ''
-
-      let novelFoodFields = this.getFields('novelFoods')
-      if (novelFoodFields.length > 0) {
-        novelFoodQueryPart += `
-        ${novelFoodFields.includes('nfCode') ? 'nfCode' : ''}
-        ${novelFoodFields.includes('title') ? 'title' : ''}
-        ${novelFoodFields.includes('opinionDocumentType') ? 'opinionDocumentType' : ''}
-        ${novelFoodFields.includes('opinionTitle') ? 'opinionTitle' : ''}
-        ${novelFoodFields.includes('opinionDoi') ? 'opinionDoi' : ''}
-        ${novelFoodFields.includes('opinionUrl') ? 'opinionUrl' : ''}
-        ${novelFoodFields.includes('opinionPublicationDate') ? 'opinionPublicationDate' : ''}
-        ${novelFoodFields.includes('opinionAdoptionDate') ? 'opinionAdoptionDate' : ''}
-        ${novelFoodFields.includes('toxStudyRequired') ? 'toxStudyRequired' : ''}
-        ${novelFoodFields.includes('genotoxFinalOutcome') ? 'genotoxFinalOutcome' : ''}
-        ${novelFoodFields.includes('finalToxicologyRemarks') ? 'finalToxicologyRemarks' : ''}
-        ${novelFoodFields.includes('proteinDigestibility') ? 'proteinDigestibility' : ''}
-        ${novelFoodFields.includes('antinutritionalFactors') ? 'antinutritionalFactors' : ''}
-        ${novelFoodFields.includes('hasNutriDisadvantage') ? 'hasNutriDisadvantage' : ''}
-        ${novelFoodFields.includes('nutriDisadvantageExplanation') ? 'nutriDisadvantageExplanation' : ''}
-        ${novelFoodFields.includes('sufficientData') ? 'sufficientData' : ''}
-        ${novelFoodFields.includes('foodMatrices') ? 'foodMatrices' : ''}
-        ${novelFoodFields.includes('instabilityConcerns') ? 'instabilityConcerns' : ''}
-        ${novelFoodFields.includes('shelflifeValue') ? 'shelflifeValue' : ''}
-        ${novelFoodFields.includes('shelflifeUnit') ? 'shelflifeUnit' : ''}
-        ${novelFoodFields.includes('endocrineDisruptProp') ? 'endocrineDisruptProp' : ''}
-        ${novelFoodFields.includes('outcome') ? 'outcome' : ''}
-        ${novelFoodFields.includes('outcomeRemarks') ? 'outcomeRemarks' : ''}
-        ${novelFoodFields.includes('vocabId') ? 'vocabId' : ''}
-        `
-      }
-
-      let panelsFields = this.getFields('panels')
-      if (panelsFields.length > 0 && panelsFields.includes('title')) {
-        novelFoodQueryPart += `
-        panels {
-          title
-        }
-        `
-      }
-
-      let sciOfficersFields = this.getFields('sciOfficers')
-      if (sciOfficersFields.length > 0) {
-        novelFoodQueryPart += `
-        sciOfficers {
-          ${sciOfficersFields.includes('firstName') ? 'firstName' : ''}
-          ${sciOfficersFields.includes('middleName') ? 'middleName' : ''}
-          ${sciOfficersFields.includes('lastName') ? 'lastName' : ''}
-        }
-        `
-      }
-
-      let allergenicityFields = this.getFields('allergenicities')
-      if (allergenicityFields.length > 0) {
-        novelFoodQueryPart += `
-        allergenicities {
-          ${allergenicityFields.includes('title') ? 'title' : ''}
-        }
-        `
-      }
-
-      let questionsFields = this.getFields('questions')
-      if (questionsFields.includes('number')) {
-        questioinsQueryPart += `
-          number
-          `
-      }
-
-      let applicantsFields = this.getFields('applicants')
-      if (applicantsFields.includes('title')) {
-        questioinsQueryPart += `
-          applicants {
-            title
-          }
-          `
-      }
-
-      let mandateFields = this.getFields('mandates')
-      if (mandateFields.length > 0) {
-        questioinsQueryPart += `
-          mandates {
-            ${mandateFields.includes('mandateTypeTitle') ? 'mandateTypeTitle' : ''}
-            ${mandateFields.includes('mandateTypeDefinition') ? 'mandateTypeDefinition' : ''}
-            ${mandateFields.includes('regulation') ? 'regulation' : ''}
-          }
-        `
-      }
-
-      if (questioinsQueryPart) {
-        novelFoodQueryPart += `
-        questions {
-          ${questioinsQueryPart}
-        }
-        `
-      }
-      const finalQuery = gql`
-      query GetNovelFoods($novelFoodTitle: String) {
-          novelFoods(novelFoodTitle: $novelFoodTitle) {
-            ${novelFoodQueryPart}
-            }
-            }`
-
-      return finalQuery
-    },
+    buildQraphQLQuery: buildQraphQLQuery,
     async renderTable(addedFilters, selectedAttrs) {
       this.addedFilters = addedFilters
       this.selectedAttrs = selectedAttrs
       this.tableIsLoading = true
-      const GET_CHOSEN_DATA = this.buildQraphQLQuery()
+      const GET_CHOSEN_DATA = this.buildQraphQLQuery(this.selectedAttrs)
       try {
-        const GET_ALL_DATA = `
-          query {
-            novelFoods {
-              nfCode
-              title
-              opinionDocumentType
-              opinionTitle
-              opinionDoi
-              opinionUrl
-              opinionPublicationDate
-              opinionAdoptionDate
-              toxStudyRequired
-              genotoxFinalOutcome
-              finalToxicologyRemarks
-              proteinDigestibility
-              antinutritionalFactors
-              hasNutriDisadvantage
-              nutriDisadvantageExplanation
-              sufficientData
-              foodMatrices
-              instabilityConcerns
-              shelflifeValue
-              shelflifeUnit
-              endocrineDisruptProp
-              outcome
-              outcomeRemarks
-              vocabId
-              allergenicity
-
-              panels {
-                title
-              }
-              sciOfficers {
-                firstName
-                middleName
-                lastName
-              }
-              questions {
-                number
-                applicants {
-                  title
-                }
-                mandates {
-                  mandateTypeTitle
-                  mandateTypeDefinition
-                  regulation
-                }
-              }
-            }
-          }
-        `
         // for Composition API apollo provider
         // const { client } = useApolloClient()
         const nfTitle = this.addedFilters.filter((filter) => filter.title === 'novel food title')[0]
